requirements.txt
google-generativeai
python-dotenv
pandas
PyPDF2
pypdf